This is project file containing all the rave gravely tunes
This is an Elektron Analog Keys/Four Mk1 file. 
It contains the songs that were used in the Rave Gravely ep:
https://thediscoparticipationmatrix.bandcamp.com/releases
It also contains songs that we toyed with but did not continue to use for various reasons.
All this is open source under the GNU general public license so feel free to use as you like.